Make LoadPlaceFile.ashx look better
For some reason, the config include was in another PHP wrapper altogether. I put it in the same wrapper as the main code, and made $pid use addslashes($_GET["'PlaceId']) instead of raw $_GET. Also added a sign() function to make signing easier. TODO: Maybe put this in a main include file?